Rename initModule -> modulesLoader and export it exclusively
