+++ test/functional/http/get/delegates.js
@@ -4,7 +4,7 @@ var node = require('../../../node.js');
 var http = require('../../../common/httpCommunication.js');
 var sendLISK = require('../../../common/complexTransactions.js').sendLISK;
 var sendTransaction = require('../../../common/complexTransactions.js').sendTransaction;
-var modulesLoader = require('../../../common/initModule.js').modulesLoader;
+var modulesLoader = require('./../common/modulesLoader');
 var genesisDelegates = require('../../../genesisDelegates.json');
 var testDelegate = genesisDelegates.delegates[0];
